# Conflicts:
#	api/lib/bad_request_error.rb

+++ public/javascripts/plugins/fileupload/multifileupload.js
@@ -341,6 +341,18 @@ Helpdesk = Helpdesk || {};
             _this.currentProgress.xhr().abort();
             return "done";
         },
+        fileAbortAll: function() {
+            var LoopInterval = setInterval(removeData,200);
+            function removeData() {
+                var _this = this;
+                 var newAttachments = $(".multiple-filelist").find('.file-upload .file-abort');
+                 newAttachments.first().click();
+                 var newLength = newAttachments.length - 1;
+                 if(newLength == 0) {
+                    clearInterval(LoopInterval);
+                 }
+            }
+        },
         // --------------------- Background proccess starts here ---- 
         backgroundProcess: {
             pool: [],
@@ -455,6 +467,8 @@ Helpdesk = Helpdesk || {};
                         if (!c) {
                             event.preventDefault();
                         } else {
+                            _this.fileAbortAll();
+                            _this.upload_status = _this.upload_status.splice(_this.upload_status.length);
                             _this.PjaxOff();
                         }
                     });
@@ -493,9 +507,7 @@ Helpdesk = Helpdesk || {};
             });
             // delete action for  - existing file upload
             $('body').on("click", ".existing-filelist .file-close , .existing-filelist-cloud .file-close", function() {
-                var c = confirm(_this.message.permanent_delete);
                 var element_this = $(this);
-                if (c) {
                     if (element_this.data('cloudFile')) {
                         var deleteUrl = window.location.origin + "/helpdesk/cloud_files/" + $(this).data('attachId');
                         if (App.namespace == "solution/articles/show") {
@@ -517,7 +529,11 @@ Helpdesk = Helpdesk || {};
                                 $("#helpdesk_attachment_" + element_this.data('attachId')).remove();
                             }
                         }
-
+                         // confirm message
+                        var c = confirm(_this.message.permanent_delete);
+                        if(!c) {
+                          return;
+                        }
                         $.ajax({
                             url: deleteUrl,
                             type: "delete",
@@ -551,6 +567,11 @@ Helpdesk = Helpdesk || {};
                                 $("#helpdesk_attachment_" + element_this.data('attachId')).remove();
                             }
                         }
+                        // confirm message
+                        var c = confirm(_this.message.permanent_delete);
+                        if(!c) {
+                          return;
+                        }
                         $.ajax({
                             url: deleteUrl,
                             type: "delete",
@@ -564,7 +585,6 @@ Helpdesk = Helpdesk || {};
                             }
                         });
                     }
-                }
 
             });
             // removing cloud files and canned response
@@ -660,27 +680,29 @@ Helpdesk = Helpdesk || {};
                     }
                 });
             });
-            // reminder
-            // jQuery(window).on("pjax:beforeSend.attachment-reminder", function(event) {
-            //     if (jQuery(".multiple-filelist .file").length !== 0 && _this.upload_status.length == 0 && jQuery(".attach-block-ui:visible").length == 0) {
-            //         if(_this.reminder) {
-            //             var c = confirm(_this.message.reminder);
-            //             if (!c) {
-            //                 event.preventDefault();
-            //                 event.stopImmediatePropagation();
-            //                 return;
-            //             }
-            //         }
-            //     }
-            // });
-            // jQuery(window).on("beforeunload", function() {
-            //     if (jQuery(".multiple-filelist .file").length !== 0) {
-            //         if(_this.reminder) {
-            //             return _this.message.reminder;
-            //         }
-            //     }
-            // });
-
+            // reminders 
+            $(window).on("pjax:beforeSend.attachment-reminder", function(event) {
+                if ($(".multiple-filelist .file").length !== 0 && _this.upload_status.length == 0 && $(".attach-block-ui:visible").length == 0) {
+                    if(_this.reminder) {
+                        var c = confirm(_this.message.reminder);
+                        if (!c) {
+                            event.preventDefault();
+                            event.stopImmediatePropagation();
+                            return;
+                        }
+                    }
+                }
+            });
+            // on window reload or close
+            $(window).on("beforeunload", function() {
+                if(_this.reminder) {
+                    if (($(".multiple-filelist .file").length !== 0) && (_this.upload_status.length == 0)) {
+                    return "clear attachments ? ";
+                }
+            }
+                
+            });
+            
         },
 
         // ** global functions **
@@ -777,9 +799,6 @@ Helpdesk = Helpdesk || {};
                 }));
                 this.manage_counts(container);
             }
-
-
-
         },
         manage_counts: function(container) {
             // managing counts
@@ -798,7 +817,7 @@ Helpdesk = Helpdesk || {};
             } else {
                 $(container).children('h5').children('strong').text('');
             }
-        }
+        },
     };
 
 }(jQuery));
